size_t change

+++ Libft/ft_memccpy.c
@@ -6,13 +6,13 @@
 /*   By: piotroff <piotroff@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/05/08 19:11:23 by piotroff          #+#    #+#             */
-/*   Updated: 2023/05/08 20:38:26 by piotroff         ###   ########.fr       */
+/*   Updated: 2023/05/15 18:30:51 by piotroff         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-void *ft_memccpy(void *dest, const void *src, int c, unsigned int n)
+void *ft_memccpy(void *dest, const void *src, int c, size_t n)
 {
     char	*d; 
 	char	*s;
