+++ Libft/ft_memchr.c
@@ -6,13 +6,13 @@
 /*   By: piotroff <piotroff@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/05/03 19:00:48 by apiotrau          #+#    #+#             */
-/*   Updated: 2023/05/08 20:38:26 by piotroff         ###   ########.fr       */
+/*   Updated: 2023/05/15 18:30:32 by piotroff         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-void *ft_memchr(const void *s, int c, unsigned int n)
+void *ft_memchr(const void *s, int c, size_t n)
 {
 	unsigned int	i;
 	char *p;
