+++ Libft/ft_strncmp.c
@@ -6,11 +6,13 @@
 /*   By: piotroff <piotroff@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/05/03 18:15:58 by apiotrau          #+#    #+#             */
-/*   Updated: 2023/05/11 19:23:44 by piotroff         ###   ########.fr       */
+/*   Updated: 2023/05/15 18:31:32 by piotroff         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-int	ft_strncmp(const char *s1, const char *s2, unsigned int n)
+#include "libft.h"
+
+int	ft_strncmp(const char *s1, const char *s2, size_t n)
 {
 	unsigned int	i;
 	unsigned char	*c1;
