+++ Libft/ft_strnstr.c
@@ -6,7 +6,7 @@
 /*   By: piotroff <piotroff@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/05/03 19:30:39 by apiotrau          #+#    #+#             */
-/*   Updated: 2023/05/11 19:25:08 by piotroff         ###   ########.fr       */
+/*   Updated: 2023/05/15 18:25:27 by piotroff         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,7 +17,7 @@ int comp_str(char c1, char c2)
 	return (c1 - c2);
 }
 
-char *ft_strnstr(const char *big, const char *little, unsigned int len)
+char *ft_strnstr(const char *big, const char *little, size_t len)
 {
 	int	i;
 	int	j;
